+++ src/lib.rs
@@ -10,8 +10,8 @@ extern crate alloc as std;
 #[cfg(feature = "std")]
 extern crate std;
 
-#[cfg(feature = "varing")]
-#[cfg_attr(docsrs, doc(cfg(feature = "varing")))]
+#[cfg(feature = "varint")]
+#[cfg_attr(docsrs, doc(cfg(feature = "varint")))]
 pub use varing::Varint;
 
 pub use chunk::*;
@@ -19,6 +19,18 @@ pub use chunk_mut::*;
 
 /// Errors buffer I/O
 pub mod error;
+
+macro_rules! non_zero {
+  ($($size:literal),+$(,)?) => {
+    paste::paste! {
+      $(
+        const [<NON_ZERO_ $size>]: ::core::num::NonZeroUsize = core::num::NonZeroUsize::new($size).expect(concat!($size, "is non-zero"));
+      )*
+    }
+  };
+}
+
+non_zero!(1);
 
 mod chunk;
 mod chunk_mut;
@@ -31,4 +43,20 @@ fn panic_advance(error_info: &error::TryAdvanceError) -> ! {
     error_info.available(),
     error_info.requested()
   );
+}
+
+/// # Panics
+/// This function panics if `size` is zero. Use only when you have already checked that `size` is non-zero.
+#[inline]
+const fn must_non_zero(size: usize) -> core::num::NonZeroUsize {
+  match core::num::NonZeroUsize::new(size) {
+    Some(nz) => nz,
+    None => panic!("Already checked value is non-zero"),
+  }
+}
+
+#[test]
+#[should_panic]
+fn test_must_non_zero_panic() {
+  must_non_zero(0);
 }